Fixed GUFuncWrapper and compare_numba_vs_python

+++ tests/processors/test_fixed_time_pickoff.py
@@ -15,57 +15,46 @@ def test_fixed_time_pickoff(compare_numba_vs_python):
     # test for nan if w_in has a nan
     w_in = np.ones(len_wf)
     w_in[4] = np.nan
-    assert np.isnan(
-        compare_numba_vs_python(fixed_time_pickoff, w_in, 1, np.int8(ord("i")))
-    )
+    assert np.isnan(compare_numba_vs_python(fixed_time_pickoff, w_in, 1, ord("i")))
 
     # test for nan if nan is passed to t_in
     w_in = np.ones(len_wf)
-    assert np.isnan(
-        compare_numba_vs_python(fixed_time_pickoff, w_in, np.nan, np.int8(ord("i")))
-    )
+    assert np.isnan(compare_numba_vs_python(fixed_time_pickoff, w_in, np.nan, ord("i")))
 
     # test for nan if t_in is negative
     w_in = np.ones(len_wf)
-    assert np.isnan(
-        compare_numba_vs_python(fixed_time_pickoff, w_in, -1, np.int8(ord("i")))
-    )
+    assert np.isnan(compare_numba_vs_python(fixed_time_pickoff, w_in, -1, ord("i")))
 
     # test for nan if t_in is too large
     w_in = np.ones(len_wf)
-    assert np.isnan(
-        compare_numba_vs_python(fixed_time_pickoff, w_in, len_wf, np.int8(ord("i")))
-    )
+    assert np.isnan(compare_numba_vs_python(fixed_time_pickoff, w_in, len_wf, ord("i")))
 
     # test for DSPFatal errors being raised
     # noninteger t_in with integer interpolation
     with pytest.raises(DSPFatal):
         w_in = np.ones(len_wf)
-        fixed_time_pickoff(w_in, 1.5, np.int8(ord("i")))
+        fixed_time_pickoff(w_in, 1.5, ord("i"))
     with pytest.raises(DSPFatal):
         a_out = np.empty(1)
-        inspect.unwrap(fixed_time_pickoff)(w_in, 1.5, np.int8(ord("i")), a_out)
+        inspect.unwrap(fixed_time_pickoff)(w_in, 1.5, ord("i"), a_out)
 
     # unsupported mode_in character
     with pytest.raises(DSPFatal):
         w_in = np.ones(len_wf)
-        fixed_time_pickoff(w_in, 1.5, np.int8(ord(" ")))
+        fixed_time_pickoff(w_in, 1.5, ord(" "))
     with pytest.raises(DSPFatal):
         a_out = np.empty(1)
-        inspect.unwrap(fixed_time_pickoff)(w_in, 1.5, np.int8(ord(" ")), a_out)
+        inspect.unwrap(fixed_time_pickoff)(w_in, 1.5, ord(" "), a_out)
 
     # linear tests
     w_in = np.arange(len_wf, dtype=float)
-    assert compare_numba_vs_python(fixed_time_pickoff, w_in, 3, np.int8(ord("i"))) == 3
+    assert compare_numba_vs_python(fixed_time_pickoff, w_in, 3, ord("i")) == 3
 
     chars = ["n", "f", "c", "l", "h", "s"]
     sols = [4, 3, 4, 3.5, 3.5, 3.5]
 
     for char, sol in zip(chars, sols):
-        assert (
-            compare_numba_vs_python(fixed_time_pickoff, w_in, 3.5, np.int8(ord(char)))
-            == sol
-        )
+        assert compare_numba_vs_python(fixed_time_pickoff, w_in, 3.5, ord(char)) == sol
 
     # sine wave tests
     w_in = np.sin(np.arange(len_wf))
@@ -82,7 +71,7 @@ def test_fixed_time_pickoff(compare_numba_vs_python):
 
     for char, sol in zip(chars, sols):
         assert np.isclose(
-            compare_numba_vs_python(fixed_time_pickoff, w_in, 3.25, np.int8(ord(char))),
+            compare_numba_vs_python(fixed_time_pickoff, w_in, 3.25, ord(char)),
             sol,
         )
 
@@ -96,6 +85,6 @@ def test_fixed_time_pickoff(compare_numba_vs_python):
 
     for ftp, sol in zip(ftps, sols):
         assert np.isclose(
-            compare_numba_vs_python(fixed_time_pickoff, w_in, ftp, np.int8(ord("h"))),
+            compare_numba_vs_python(fixed_time_pickoff, w_in, ftp, ord("h")),
             sol,
         )
