+++ tests/processors/test_time_point_thresh.py
@@ -93,16 +93,14 @@ def test_interpolated_time_point_thresh(compare_numba_vs_python):
     w_in = np.concatenate([np.arange(-1, 5, 1), np.arange(-1, 5, 1)], dtype="float")
     w_in[4] = np.nan
     assert np.isnan(
-        compare_numba_vs_python(
-            interpolated_time_point_thresh, w_in, 1.0, 11.0, 0, np.int8(105)
-        )
+        compare_numba_vs_python(interpolated_time_point_thresh, w_in, 1.0, 11.0, 0, 105)
     )
 
     # test for nan if nan is passed to a_threshold
     w_in = np.concatenate([np.arange(-1, 5, 1), np.arange(-1, 5, 1)], dtype="float")
     assert np.isnan(
         compare_numba_vs_python(
-            interpolated_time_point_thresh, w_in, np.nan, 11.0, 0, np.int8(105)
+            interpolated_time_point_thresh, w_in, np.nan, 11.0, 0, 105
         )
     )
 
@@ -110,105 +108,83 @@ def test_interpolated_time_point_thresh(compare_numba_vs_python):
     w_in = np.concatenate([np.arange(-1, 5, 1), np.arange(-1, 5, 1)], dtype="float")
     assert np.isnan(
         compare_numba_vs_python(
-            interpolated_time_point_thresh, w_in, 1.0, np.nan, 0, np.int8(105)
+            interpolated_time_point_thresh, w_in, 1.0, np.nan, 0, 105
         )
     )
 
     # test for nan if t_start out of range
     w_in = np.concatenate([np.arange(-1, 5, 1), np.arange(-1, 5, 1)], dtype="float")
     assert np.isnan(
-        compare_numba_vs_python(
-            interpolated_time_point_thresh, w_in, 1.0, 12, 0, np.int8(105)
-        )
+        compare_numba_vs_python(interpolated_time_point_thresh, w_in, 1.0, 12, 0, 105)
     )
 
     # test walk backward mode 'i'
     w_in = np.concatenate([np.arange(-1, 5, 1), np.arange(-1, 5, 1)], dtype="float")
     assert (
-        compare_numba_vs_python(
-            interpolated_time_point_thresh, w_in, 1, 11, 0, np.int8(105)
-        )
+        compare_numba_vs_python(interpolated_time_point_thresh, w_in, 1, 11, 0, 105)
         == 7.0
     )
 
     # test walk forward mode 'i'
     w_in = np.concatenate([np.arange(-1, 5, 1), np.arange(-1, 5, 1)], dtype="float")
     assert (
-        compare_numba_vs_python(
-            interpolated_time_point_thresh, w_in, 3, 0, 1, np.int8(105)
-        )
+        compare_numba_vs_python(interpolated_time_point_thresh, w_in, 3, 0, 1, 105)
         == 4.0
     )
 
     # test walk backward mode 'f'
     w_in = np.concatenate([np.arange(-1, 5, 1), np.arange(-1, 5, 1)], dtype="float")
     assert (
-        compare_numba_vs_python(
-            interpolated_time_point_thresh, w_in, 1, 11, 0, np.int8(102)
-        )
+        compare_numba_vs_python(interpolated_time_point_thresh, w_in, 1, 11, 0, 102)
         == 8.0
     )
 
     # test walk forward mode 'f'
     w_in = np.concatenate([np.arange(-1, 5, 1), np.arange(-1, 5, 1)], dtype="float")
     assert (
-        compare_numba_vs_python(
-            interpolated_time_point_thresh, w_in, 3, 0, 1, np.int8(102)
-        )
+        compare_numba_vs_python(interpolated_time_point_thresh, w_in, 3, 0, 1, 102)
         == 5.0
     )
 
     # test walk backward mode 'f'
     w_in = np.concatenate([np.arange(-1, 5, 1), np.arange(-1, 5, 1)], dtype="float")
     assert (
-        compare_numba_vs_python(
-            interpolated_time_point_thresh, w_in, 1, 11, 0, np.int8(99)
-        )
+        compare_numba_vs_python(interpolated_time_point_thresh, w_in, 1, 11, 0, 99)
         == 7.0
     )
 
     # test walk forward mode 'f'
     w_in = np.concatenate([np.arange(-1, 5, 1), np.arange(-1, 5, 1)], dtype="float")
     assert (
-        compare_numba_vs_python(
-            interpolated_time_point_thresh, w_in, 3, 0, 1, np.int8(99)
-        )
+        compare_numba_vs_python(interpolated_time_point_thresh, w_in, 3, 0, 1, 99)
         == 4.0
     )
 
     # test walk backward mode 'n'
     w_in = np.concatenate([np.arange(-1, 5, 1), np.arange(-1, 5, 1)], dtype="float")
     assert (
-        compare_numba_vs_python(
-            interpolated_time_point_thresh, w_in, 1, 11, 0, np.int8(110)
-        )
+        compare_numba_vs_python(interpolated_time_point_thresh, w_in, 1, 11, 0, 110)
         == 7.5
     )
 
     # test walk forward mode 'n'
     w_in = np.concatenate([np.arange(-1, 5, 1), np.arange(-1, 5, 1)], dtype="float")
     assert (
-        compare_numba_vs_python(
-            interpolated_time_point_thresh, w_in, 3, 0, 1, np.int8(110)
-        )
+        compare_numba_vs_python(interpolated_time_point_thresh, w_in, 3, 0, 1, 110)
         == 4.5
     )
 
     # test walk backward mode 'l'
     w_in = np.concatenate([np.arange(-1, 5, 1), np.arange(-1, 5, 1)], dtype="float")
     assert (
-        compare_numba_vs_python(
-            interpolated_time_point_thresh, w_in, 1.5, 11, 0, np.int8(108)
-        )
+        compare_numba_vs_python(interpolated_time_point_thresh, w_in, 1.5, 11, 0, 108)
         == 8.5
     )
 
     # test walk forward mode 'l'
     w_in = np.concatenate([np.arange(-1, 5, 1), np.arange(-1, 5, 1)], dtype="float")
     assert (
-        compare_numba_vs_python(
-            interpolated_time_point_thresh, w_in, 3.5, 0, 1, np.int8(108)
-        )
+        compare_numba_vs_python(interpolated_time_point_thresh, w_in, 3.5, 0, 1, 108)
         == 4.5
     )
